==> wd_adni/script02_merge_subject_info.ipynb <==
## merge subject ids in the test set with demographic info.

## updates on Jun 28:
## add MEM and EF (for ML predition) score to subject info.
## use baseline fMRI features to predict further MEM and EF scores.

import pandas as pd
import os
import re

input_file   = 'adni_out01_time_signal_test_train_split/test_set_roi_signals.csv'
input_file2  = 'adni_subject_info/MRI_information_All_Measurement.csv'
input_file3  = 'adni_subject_info/ADNI_MEM_EF_Composit_Score.csv'

output_dir   = 'adni_out02_test_subject_info'

file_test    = pd.read_csv(input_file, header = None)
file_test    = file_test[0].tolist()
subject_test = set()

subject_id = []
ses_id = []

for f in file_test:
    subject_id += [re.search('(.*)_sub-(.*)-ses-(.*).txt', f).group(2)]
    ses_id +=  [re.search('(.*)_sub-(.*)-ses-(.*).txt', f).group(3)]

subject_info = pd.DataFrame({'subject_id': subject_id, 'ses_id': ses_id})


subject_info

demographic_data = pd.read_csv(input_file2)
g = demographic_data.groupby(['PTID']).first().reset_index()
df_merge = subject_info.merge(g[['PTID', 'AGE', 'PTGENDER', 'DX_bl', 'DX']], how = 'left', 
                              left_on = 'subject_id', right_on = 'PTID').drop(columns = 'PTID')
df_merge

if not os.path.isdir(output_dir):
    os.mkdir(output_dir)

df_merge.to_csv(output_dir + '/subject_info_merge.csv')


composite_score = pd.read_csv(input_file3)
composite_score = composite_score.dropna(subset=['ADNI_MEM', 'ADNI_EF'])
composite_score = composite_score.sort_values('EXAMDATE', ascending=False).drop_duplicates(['PTID'])
composite_score.sort_values('PTID')

df_merge2 = df_merge.merge(composite_score[['PTID', 'EXAMDATE', 'ADNI_MEM', 'ADNI_EF']], how = 'left', 
                              left_on = 'subject_id', right_on = 'PTID').drop(columns = 'PTID')
df_merge2

df_merge2.to_csv(output_dir + '/subject_info_merge2.csv')
