==> wd_adni/script05_lasso_on_bold_entropy.ipynb <==
## run LASSO on the entropy of orignal bold signals.

import pandas as pd
import numpy as np
import os
import re
import glob

from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LogisticRegressionCV
from sklearn.model_selection import cross_val_score

input_dir = 'adni_out04_bold_entropy/'
output_dir = 'adni_out05_lasso_bold_entropy/'

## subject info:
input_subject_info = 'adni_out02_test_subject_info/subject_info_merge.csv'
subject_info = pd.read_csv(input_subject_info, index_col = 0)
subject_info.sort_values(by = ['subject_id', 'ses_id'], inplace = True)
subject_info


# data_files = glob.glob(input_dir + '/adni_out04_entropy_freq_*.csv')
metrics = ['adni_out04_bold_entropy']
data_files = [input_dir + m + '.csv' for m in metrics ]
data_files.sort()

all_result = []
print_group_size = True
# comparision = [('CN', 'AD'), ('CN', 'EMCI'), ('CN', 'LMCI'), ('CN', 'SMC')]
comparision = [('CN', 'AD'), ('CN', 'EMCI'), ('CN', 'LMCI')]
result_accuracy = {comp: [] for comp in comparision}
result_coefs = {comp: [] for comp in comparision}

for f in data_files:
    
    print(f)
    data = pd.read_csv(f, index_col = 0)
    data = subject_info.merge(data, on = ['subject_id', 'ses_id', 'AGE', 'PTGENDER', 'DX_bl', 'DX'], how = 'inner')
    data.dropna(subset=['DX'], inplace = True)
    data.drop(columns = ['ses_id', 'DX_bl'], inplace = True)
    # break
    
    if print_group_size:
        print(data.groupby(['PTGENDER', 'DX']).size())
        print_group_size = False
    
    X = pd.concat([data['AGE'], data['PTGENDER'].map({'Male': 1, 'Female': 0}), data.iloc[:, 4:]], axis = 1).values
    y = data['DX'].map({'CN': 0, 'EMCI': 1, 'LMCI': 1, 'AD': 1, 'SMC': 1}).values
    
    scaler = StandardScaler()
    scaler.fit(X)
    X = scaler.transform(X)
    C_values = np.logspace(-1, 2, 10)
    
    for comp in comparision:
        # select data with ad/lmci and cn:    
        idx_ad_cn = data['DX'].isin(comp)
        # build lassoCV that tune parameters with inner cv:
        reg = LogisticRegressionCV(cv=5, random_state=0, Cs=C_values, n_jobs = 6,
                                   penalty = 'l1', solver='liblinear', 
                                   refit = False).fit(X[idx_ad_cn,:], y[idx_ad_cn])
        # reg = LassoCV(cv=5, random_state=0, normalize = True, tol = .001).fit(X[idx_ad_cn,:], y[idx_ad_cn])
        # run outer cv:
        cv_result = cross_val_score(reg, X[idx_ad_cn,:], y[idx_ad_cn], cv=5)
        print('cross validaton result, mean %3f, std: %3f' % (cv_result.mean(), cv_result.std()))
        result_accuracy[comp]=list(cv_result)
        result_coefs[comp]=list(reg.coef_.reshape(-1))
        
    #     break
    # break
    
print('finished')


## save results:
## save all accuracy results to one file (as we only have 1 metric for each comparision)
## udpates on May-10-2022.

if not os.path.exists(output_dir):
    os.mkdir(output_dir)

res = pd.DataFrame(result_accuracy)
res.index = ['cv' + str(i) for i in range(1, 6)]
res.columns = ['CV vs. AD', 'CN vs. EMCI', 'CN vs. LMCI']

res.to_csv(output_dir + '/lasso_accuracy.csv')


res = pd.DataFrame(result_coefs)
res.index = ['Age', 'Sex'] + ['roi' + str(i) for i in range(1, 265)]
res.columns = ['CV vs. AD', 'CN vs. EMCI', 'CN vs. LMCI']

res.to_csv(output_dir + '/lasso_coefs.csv')


res
# result_accuracy

metrics
