==> wd_adni/script05_ML_predict_adni_score.ipynb <==
# predict adni MEM and EF score based on brain imaging features:
## 1. harmonic entropy
## 2. network metrics

import pandas as pd
import numpy as np
import os
import re
import glob

from sklearn.preprocessing import StandardScaler
# from sklearn.linear_model import LogisticRegressionCV
from sklearn.linear_model import RidgeCV
from sklearn.metrics import r2_score

input_dir = 'adni_out04_power_histogram_and_entropy'
output_dir = 'adni_out05_composite_score_prediction_entropy'

if not os.path.exists(output_dir):
    os.mkdir(output_dir)


from sklearn.model_selection import KFold
# from sklearn.model_selection import StratifiedKFold

# tune model in n-fold CV:
# n fold CV on hc data:
nfold = 3
seed = 123
kf = KFold(n_splits=nfold, shuffle = True, random_state=seed)

n_alphas = 100
alphas = np.logspace(0, 4, n_alphas)
ridge_cv = RidgeCV(alphas=alphas, cv = 3)

## SVR with grid_search:
from sklearn.svm import SVR
from sklearn.model_selection import GridSearchCV
 
# defining parameter range
param_grid = {'C': [.01, 0.1, 1, 10],
              'gamma': [5, 1, 0.1, 0.01, 0.001, 0.0001],
              'epsilon': [.025, .05, .1, .25, .5],
              'kernel': ['rbf', 'sigmoid']}
 
svr_grid = GridSearchCV(SVR(), param_grid, refit = True, verbose = True, n_jobs = 3, cv = 3)

data_files = glob.glob(input_dir + '/adni_out04_entropy_freq_*.csv')
data_files.sort()

input_subject_info = 'adni_out02_test_subject_info/subject_info_merge2.csv'
subject_info = pd.read_csv(input_subject_info, index_col = 0)

column_index = pd.MultiIndex.from_product([['Pearson r', 'R square', 'MAE', 'rmse'],
                                           ['cv' + str(i) for i in range(1, kf.n_splits+1, 1)]])

row_index = ['freq' + str(i) for i in range(1, len(data_files) + 1)]
result_table = pd.DataFrame(index = row_index, columns = column_index)

response = ['ADNI_MEM', 'ADNI_EF']
result_table_all = {resp: result_table.copy() for resp in response}

for i_freq, f in enumerate(data_files):
    
    plot_data = pd.DataFrame(columns = ['SUBJID', 'CV'])
    
    print(f)
    freq = re.search('(.*)_freq_(.*).csv', f).group(2)
    data = pd.read_csv(f, index_col = 0)
    
    data = subject_info.merge(data.iloc[:, [0] + [i for i in range(6, data.shape[1])]], on = 'subject_id')
    data.dropna(subset=response, inplace = True)
    subjid = data['subject_id'].values
    
    X = pd.concat([data['PTGENDER'].map({'Male': 1, 'Female': 0}), 
                   data['AGE'], 
                   data.iloc[:, 8:]], axis = 1).values
    
    i = 1
    for train_index, test_index in kf.split(X):
        print('run_model on CV: %d' % i)

        X_train, X_test = X[train_index,:], X[test_index,:]
        y_train, y_test = y[train_index], y[test_index]

        # normalize X_train and X_test based on mean and sd of X_train. 
        # don't normalize 1st col which is sex, 
        scaler = StandardScaler().fit(X_train[:,1:])
        X_train[:,1:] = scaler.transform(X_train[:,1:])  
        X_test[:,1:] = scaler.transform(X_test[:,1:])
        subjid_test = subjid[test_index] 
        
        seed = 7
        np.random.seed(seed)           
    
        plot_data_cv = pd.DataFrame(columns = ['SUBJID', 'CV'])
        for resp in response:

            y = data[resp].values
            y_train = y[train_index]
            y_test  = y[test_index]
            
            
            fit_result = ridge_cv.fit(X_train, y_train)
            y_prediction = ridge_cv.predict(X_test)
            print(ridge_cv.alpha_)
            
            # fitting the model for grid search
            # svr_grid.fit(X_train, y_train)
            # print(svr_grid.best_params_)
            # y_prediction = svr_grid.predict(X_test)

            result_table = result_table_all[resp]
            result_table['R square', 'cv' + str(i)][i_freq] = r2_score(y_test, y_prediction)
            result_table['Pearson r', 'cv' + str(i)][i_freq] = np.corrcoef(y_test, y_prediction)[0,1]
            result_table['rmse', 'cv' + str(i)][i_freq] = np.sqrt(np.mean(np.square(y_test - y_prediction)))
            result_table['MAE', 'cv' + str(i)][i_freq] = np.mean(np.abs(y_test - y_prediction))

            plot_data_cv[resp] = y_test
            plot_data_cv[resp + '_pred'] = y_prediction
            plot_data_cv.loc[:,'CV'] = i
            plot_data_cv['SUBJID'] = subjid_test
            plot_data_cv['Sex'] = X_test[:,0]

        plot_data = pd.concat([plot_data,plot_data_cv])    
        plot_data.to_csv(output_dir+"/scatter_data_ridge_freq" + str(i_freq) + ".csv")
        # plot_data.to_csv(output_dir+"/scatter_data_svr_freq" + str(i_freq) + ".csv")
            
        i += 1
        

print('finished')

result_table

def summary_result(result_table):
    import pandas as pd
    
    column_index = pd.MultiIndex.from_product([['Pearson r', 'R square', 'MAE', 'rmse'], \
                                               ['mean', 'std']])
    
    result_table_summary = pd.DataFrame(columns = column_index)

    result_table_summary['Pearson r', 'mean'] = result_table['Pearson r'].mean(axis = 1)
    result_table_summary['Pearson r', 'std'] = result_table['Pearson r'].std(axis = 1)

    result_table_summary['R square', 'mean'] = result_table['R square'].mean(axis = 1)
    result_table_summary['R square', 'std'] = result_table['R square'].std(axis = 1)
    
    result_table_summary['MAE', 'mean'] = result_table['MAE'].mean(axis = 1)
    result_table_summary['MAE', 'std'] = result_table['MAE'].std(axis = 1)

    result_table_summary['rmse', 'mean'] = result_table['rmse'].mean(axis = 1)
    result_table_summary['rmse', 'std'] = result_table['rmse'].std(axis = 1)
    
    #print('summary_result:')
    #print(result_table_summary)
    
    result_table2 = result_table[['Pearson r', 'R square']].reset_index(level = 0, inplace = False)
    result_table3 = result_table2.rename(columns ={'index':'frequency'}, inplace = False)
    result_accuracy_plot = pd.melt(result_table3, id_vars=['frequency'], var_name = "cv")
    
#     result_table2 = result_table['R square'].reset_index(level = 0, inplace = False)
#     result_table3 = result_table2.rename(columns ={'index':'feature'}, inplace = False)
#     result_r2_plot = pd.melt(result_table3, id_vars=['feature'], \
#                                    value_name = "R square", var_name = "boot")
    #print(result_accuracy_plot)
    return((result_table_summary, result_accuracy_plot))

result_table_summary, result_accuracy_plot = summary_result(result_table)

# result_table_summary.to_csv(output_dir + '/result_table_svr.csv')
result_table_summary.to_csv(output_dir + '/result_table_ridge.csv')
result_table_summary
